Return the proper headers on errors before WebSocket handshake.
Before when there was, for example, a PermissionDenied response in
WS4REDIS_PROCESS_REQUEST, the response would only be the text of the
exception, without the standard HTTP headers.
When proxying uwsgi through nginx this lead to problems, as nginx
didn't pass the response with missing headers onto the client.
Now, this is safe in a production environment.

Conflicts:
	ws4redis/wsgi_server.py